add duplex messaging test file
