+++ middleware/echo.go
@@ -4,11 +4,9 @@ import "github.com/neptulon/neptulon"
 
 // Echo sends incoming messages back as is.
 func Echo(ctx *neptulon.ReqCtx) error {
-	var msg interface{}
-	if err := ctx.Params(&msg); err != nil {
+	// unmarshall incoming message into response directly
+	if err := ctx.Params(&ctx.Res); err != nil {
 		return err
 	}
-
-	ctx.Res = msg
 	return ctx.Next()
 }
